requirements.txt
django>=5.0.1,<6.0.0 # 5.0.1 이상 6.0.0 미만 Framework
djangorestframework>=3.14.0,<4.0.0 # DRF : For REST API
psycopg2>=2.9.9,<3.0 # Psycopg2 : For PostgreSQL, connector
drf-spectacular>=0.27.2,<0.28.0 # DFR-Spectacular : For API Documentation
